+++ PSW_Schedule_API/Services/IPSWScheduleServices.cs
@@ -6,14 +6,14 @@ namespace PSW_Schedule_API.Services
     {
         // Employee Services
         Task<List<Employee>> GetEmployeesAsync();//Get all
-        Task<Employee> GetEmployeeAsync(int id, bool includeSchedule = false); //getOne
+        Task<Employee> GetEmployeeAsync(long id, bool includeSchedule = false); //getOne
         Task<Employee> AddEmployeeAsync(Employee employee);//POST
         Task<Employee> UpdateEmployeeAsync(Employee employee);//PUT
         Task<(bool, string)> DeleteEmployeeAsync(Employee employee);//DELETE
 
         // Schedule Services
         Task<List<Schedule>> GetSchedulesAsync();//GET ALL
-        Task<Schedule> GetScheduleAsync(int id);//GET ONE
+        Task<Schedule> GetScheduleAsync(long id);//GET ONE
         Task<Schedule> AddScheduleAsync(Schedule schedule); //POST
         Task<Schedule> UpdateScheduleAsync(Schedule schedule);//PUT
         Task<(bool, string)> DeleteScheduleAsync(Schedule schedule);//DELETE
